Fix test failure, so that fn-call exprs parse
With the way the grammar is currently set up, we can't move the
`ident ( args )` to be `( ident args )`. That might be okay.

We'll probably want to fix this, though: it's helpful to have a state
transition that tells the systems "actually this is a function call"
before we have to decide whether to try to push `ident`'s double value
or otherwise call it.