Sketch out a frontend "IR"
Unsure if "IR" is the correct term. The idea is to model computation as
a register machine.

I think the type system can similarly be built as a register machine,
with instructions to (algebraically) compose types, e.g. by doing loose
sum and products, with other operations to ascribe field names / proper
names or formalize as an actual type vs. just a "temporary" type.